requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2 
certifi==2024.8.30
charset-normalizer==3.4.0
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.1
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
requests-html==0.10.0
lxml[html_clean]==4.9.3  
urllib3==2.2.3
Werkzeug==3.0.4
honcho==1.0.1
pymongo==4.10.1
